requirements.txt
Flask==2.2.2
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
requests==2.31.0
